+++ src/SWP/Bundle/ContentBundle/EventListener/ProcessArticleMediaListener.php
@@ -28,6 +28,26 @@ use SWP\Component\MultiTenancy\PathBuilder\TenantAwarePathBuilderInterface;
 class ProcessArticleMediaListener
 {
     /**
+     * @var ObjectManager
+     */
+    protected $objectManager;
+
+    /**
+     * @var TenantAwarePathBuilderInterface
+     */
+    protected $pathBuilder;
+
+    /**
+     * @var string
+     */
+    protected $mediaBasepath;
+
+    /**
+     * @var MediaManagerInterface
+     */
+    protected $mediaManager;
+
+    /**
      * ProcessArticleMediaListener constructor.
      *
      * @param ObjectManager                   $objectManager
@@ -145,6 +165,12 @@ class ProcessArticleMediaListener
         return $articleMedia;
     }
 
+    /**
+     * @param string $nodeName
+     * @param mixed  $parent
+     *
+     * @return Generic
+     */
     private function createGenericDocument($nodeName, $parent)
     {
         $document = new Generic();